feat: refactor explorer API routes and enhance explorer handling with updated methods
